+++ docs/conf.py
@@ -198,70 +198,11 @@ else:
 # Output file base name for HTML help builder.
 htmlhelp_basename = "MetalK8sdoc"
 
-
-# -- Options for LaTeX output ------------------------------------------------
-
-latex_elements = {
-    # The paper size ('letterpaper' or 'a4paper').
-    #
-    "papersize": "a4paper",
-    # The font size ('10pt', '11pt' or '12pt').
-    #
-    "pointsize": "10pt",
-    # Additional stuff for the LaTeX preamble.
-    #
-    "preamble": r"""
-        \usepackage{charter}
-        \usepackage[defaultsans]{lato}
-        \usepackage{inconsolata}
-    """,
-    # Latex figure (float) alignment
-    #
-    # 'figure_align': 'htbp',
-}
-
-# Grouping the document tree into LaTeX files. List of tuples
-# (source start file, target name, title,
-#  author, documentclass [howto, manual, or own class]).
-latex_documents = [
-    (
-        "{}-latex".format(master_doc),
-        "MetalK8s.tex",
-        "MetalK8s Documentation",
-        "Scality",
-        "manual",
-        True,
-    ),
-]
-
-latex_logo = "../artwork/generated/metalk8s-logo-wide-black.pdf"
-
-latex_toplevel_sectioning = "part"
-
-
 # -- Options for manual page output ------------------------------------------
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [(master_doc, "metalk8s", "MetalK8s Documentation", [author], 1)]
-
-
-# -- Options for Texinfo output ----------------------------------------------
-
-# Grouping the document tree into Texinfo files. List of tuples
-# (source start file, target name, title, author,
-#  dir menu entry, description, category)
-texinfo_documents = [
-    (
-        master_doc,
-        "MetalK8s",
-        "MetalK8s Documentation",
-        author,
-        "MetalK8s",
-        "One line description of project.",
-        "Miscellaneous",
-    ),
-]
 
 
 # -- Extension configuration -------------------------------------------------